# [perf] Dont refresh threat on death