@Ragnarok77-factorio 2023-10-18
We should test 394 first, and if it's not good enough, try this one, as this one is a lot more extreme (maybe too extreme, maybe other one is good enough and no need for that, even if I believe this one should improve perf more... but still..) , let's see over time.