@Ragnarok77-factorio 2023-10-25
I think 394 is enough, closing it, unless we notice it's not as good as we wanted, or feel free to open it if you gonna rework it and want it merged.